PurchaseCommand completed

+++ src/main/java/Model/Logs.java
@@ -5,9 +5,9 @@ import java.util.Date;
 public class Logs {
     protected String logID;
     protected Date date;
-    protected int pricePaid;
+    protected float pricePaid;
 
-    public Logs(String logID, Date date, int pricePaid) {
+    public Logs(String logID, Date date, float pricePaid) {
         this.logID = logID;
         this.date = date;
         this.pricePaid = pricePaid;
