+++ src/main/java/Model/SellLog.java
@@ -3,12 +3,12 @@ package Model;
 import java.util.Date;
 
 public class SellLog extends Logs {
-    private int amountReducedForOff;
+    private float amountReducedForOff;
     private Good soldGood;
     private String buyerUserNmae;
     private String postStatus;
 
-    public SellLog(String logID, Date date, int pricePaid, int amountReducedForOff, Good soldGood, String buyerUserNmae, String postStatus) {
+    public SellLog(String logID, Date date, float pricePaid, float amountReducedForOff, Good soldGood, String buyerUserNmae, String postStatus) {
         super(logID, date, pricePaid);
         this.amountReducedForOff = amountReducedForOff;
         this.soldGood = soldGood;
